Improve error messages for report with invalid input
While I was looking into MOD-184, I found that the database actually *does* check whether a report is using the right type! However, it gives a very unhelpful error message:
```json
{
  "error": "database_error",
  "description": "Database Error: error returned from database: insert or update on table \"reports\" violates foreign key constraint \"reports_mod_id_fkey\""
}
```

Thus, this PR makes it so that the error message is more friendly:
```json
{
  "error": "invalid_input",
  "description": "User could not be found: test"
}
```

I also merged the API v1 `POST /report` route into the v2 route so I didn't have to duplicate stuff and so that there wouldn't be conflicts with #502.

I'm not actually sure where the original issue with MOD-184 originates from. I wasn't able to reproduce the issue, which is even more confusing because I remember observing it happening quite often. It hasn't happened since rewrite-parity released, so I'm going to assume that the original issue is effectively fixed. In any event, this should prevent MOD-184 from happening again in case the database protection there messes up for whatever reason, and also makes the error message friendlier as an added bonus.